geolocation-sensor/GeolocationSensor-disabled-by-permissions-policy.https.html

<!DOCTYPE html>
<body>
<title>GeolocationSensor Permissions Policy Test: Disabled</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/permissions-policy/resources/permissions-policy.js"></script>
<script src="/generic-sensor/generic-sensor-feature-policy-test.sub.js"></script>
<script>
"use strict";

run_permissions_policy_tests_disabled('GeolocationSensor');

promise_test(async t => {
  await promise_rejects_dom(t, 'SecurityError', GeolocationSensor.read());
}, "GeolocationSensor.read(): 'SecurityError' is thrown when disabled by Permissions Policy");

promise_test(async t => {
  const controller = new AbortController();
  const signal = controller.signal;
  controller.abort();

  await promise_rejects_dom(t, 'AbortError', GeolocationSensor.read({ signal }));
}, "GeolocationSensor.read(): 'AbortError' takes priority");
</script>
</body>
